[CAS-1496] Fix attachment previews in pinned message list
