+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/extensions/internal/Channel.kt
@@ -3,9 +3,7 @@ package io.getstream.chat.android.ui.common.extensions.internal
 import android.content.Context
 import android.text.SpannableString
 import android.text.SpannableStringBuilder
-import com.getstream.sdk.chat.model.ModelType
 import io.getstream.chat.android.client.extensions.getUsersExcludingCurrent
-import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.livedata.ChatDomain
@@ -47,22 +45,9 @@ internal fun Channel.getLastMessagePreviewText(
             val previewText: SpannableString =
                 message.text.trim().bold(currentUserMention?.singletonList(), ignoreCase = true)
 
-            val attachments: SpannableString? = message.attachments
-                .takeIf { it.isNotEmpty() }
-                ?.mapNotNull { attachment ->
-                    attachment.title?.let { title ->
-                        val prefix = getAttachmentPrefix(attachment)
-                        if (prefix != null) {
-                            "$prefix $title"
-                        } else {
-                            title
-                        }
-                    } ?: attachment.name
-                }
-                ?.joinToString()
-                ?.italicize()
+            val attachmentsText: SpannableString? = message.getAttachmentsText()
 
-            listOf(sender, previewText, attachments)
+            listOf(sender, previewText, attachmentsText)
                 .filterNot { it.isNullOrEmpty() }
                 .joinTo(SpannableStringBuilder(), ": ")
         }
@@ -75,10 +60,4 @@ internal var Channel.isMuted: Boolean
     get() = extraData[EXTRA_DATA_MUTED] as Boolean? ?: false
     set(value) {
         extraData[EXTRA_DATA_MUTED] = value
-    }
-
-private fun getAttachmentPrefix(attachment: Attachment): String? =
-    when (attachment.type) {
-        ModelType.attach_giphy -> "/giphy"
-        else -> null
     }
