+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/extensions/internal/Message.kt
@@ -1,6 +1,9 @@
 package io.getstream.chat.android.ui.common.extensions.internal
 
 import android.content.Context
+import android.text.SpannableString
+import com.getstream.sdk.chat.model.ModelType
+import io.getstream.chat.android.client.models.Attachment
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.ui.R
 
@@ -21,3 +24,28 @@ internal fun Message.getPinnedText(context: Context): String? {
     }
     return context.getString(R.string.stream_ui_message_list_pinned_message, user)
 }
+
+/**
+ * Returns a string representation of message attachments or null if the attachment list is empty.
+ */
+internal fun Message.getAttachmentsText(): SpannableString? {
+    return attachments.takeIf { it.isNotEmpty() }
+        ?.mapNotNull { attachment ->
+            attachment.title?.let { title ->
+                val prefix = getAttachmentPrefix(attachment)
+                if (prefix != null) {
+                    "$prefix $title"
+                } else {
+                    title
+                }
+            } ?: attachment.name
+        }
+        ?.joinToString()
+        ?.italicize()
+}
+
+private fun getAttachmentPrefix(attachment: Attachment): String? =
+    when (attachment.type) {
+        ModelType.attach_giphy -> "/giphy"
+        else -> null
+    }
