+++ stream-chat-android-ui-components/src/main/res/values-hi/strings_common.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
   <string name="stream_ui_mention">@%1$s</string>
-  <string name="stream_ui_message_preview_file">"फ़ाइल: %1$s"</string>
   <string name="stream_ui_message_preview_sender">"&lt;b&gt;%2$s&lt;/b&gt; में &lt;b&gt;%1$s&lt;/b&gt;"</string>
   <string name="stream_ui_user_status_last_seen">"पिछली बार देखा %1$s"</string>
   <string name="stream_ui_user_status_last_seen_just_now">"अंतिम बार अभी देखा"</string>
