+++ stream-chat-android-ui-components/src/main/res/values-id/strings_common.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
-  <string name="stream_ui_message_preview_file">"Berkas: %1$s"</string>
   <string name="stream_ui_message_preview_sender">"&lt;b&gt;%1$s&lt;/b&gt; di &lt;b&gt;%2$s&lt;/b&gt;"</string>
   <string name="stream_ui_user_status_last_seen">"Terakhir dilihat %1$s"</string>
   <string name="stream_ui_user_status_last_seen_just_now">"Terakhir dilihat baru-baru ini"</string>
