+++ stream-chat-android-ui-components/src/main/res/values-ja/strings_common.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="stream_ui_mention">@%s</string>
-    <string name="stream_ui_message_preview_file">"ファイル: %s"</string>
     <string name="stream_ui_message_preview_sender">"&lt;b&gt;%1$s&lt;/b&gt;に&lt;b&gt;%2$s&lt;/b&gt;"</string>
     <string name="stream_ui_user_status_last_seen">"%s見ました"</string>
     <string name="stream_ui_user_status_last_seen_just_now">"今見ました"</string>
