+++ stream-chat-android-ui-components/src/main/res/values-ko/strings_common.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:xliff="urn:oasis:names:tc:xliff:document:1.2">
     <string name="stream_ui_mention">@%s</string>
-    <string name="stream_ui_message_preview_file">"파일: %s"</string>
     <string name="stream_ui_message_preview_sender">"&lt;b&gt;%1$s&lt;/b&gt;에&lt;b&gt;%2$s&lt;/b&gt;"</string>
     <string name="stream_ui_user_status_last_seen">"마지막 온라인: %s"</string>
     <string name="stream_ui_user_status_last_seen_just_now">"방금 봤습니다"</string>
